Code refactor & update units test
